+++ htsget-search/src/htsget/bcf_search.rs
@@ -155,8 +155,8 @@ mod tests {
     with_local_storage_tmp as with_local_storage_tmp_path,
   };
   use crate::htsget::{Class, Headers, Response, Url};
+  use crate::storage::data_server::HttpTicketFormatter;
   use crate::storage::local::LocalStorage;
-  use crate::storage::ticket_server::HttpTicketFormatter;
 
   use super::*;
