+++ htsget-search/src/storage/mod.rs
@@ -18,8 +18,8 @@ use crate::htsget::{Class, Headers, Url};
 
 #[cfg(feature = "s3-storage")]
 pub mod aws;
+pub mod data_server;
 pub mod local;
-pub mod ticket_server;
 
 type Result<T> = core::result::Result<T, StorageError>;
 
@@ -71,7 +71,7 @@ pub enum StorageError {
   IoError(String, io::Error),
 
   #[error("url response ticket server error: {0}")]
-  TicketServerError(String),
+  DataServerError(String),
 
   #[error("invalid input: {0}")]
   InvalidInput(String),
@@ -352,8 +352,8 @@ mod tests {
   use std::collections::HashMap;
 
   use crate::htsget::Class;
+  use crate::storage::data_server::HttpTicketFormatter;
   use crate::storage::local::LocalStorage;
-  use crate::storage::ticket_server::HttpTicketFormatter;
 
   use super::*;
